requirements.txt
gradio
pillow
transformers
torch
requests
pandas
